fix: payment entity 변경

+++ choreography/payment/src/main/java/com/example/entity/Payment.java
@@ -15,9 +15,9 @@ public class Payment {
     private int id;
 
     @Column
-    private int orderId;
+    private long orderId;
 
     @Column
-    private int amount;
+    private long amount;
 
 }
